Fix setup of local backends
The boot.looplvm init script has to be created before we create the
volume groups for local files. Also fix invalid bits in the script.

But we need to create the local files first!

And we also need to use :start instead of :reload for the boot.looplvm
notification, since :reload won't work (status never returns 0 for this
init script, since it's not a daemon that is running).